Move resolveCallable() to a trait
